Add ability to combine includes/fields
This was an initial attempt to allow the specification of multiple
filters. This would, for example, allow the detection of
duplicate requests.

Unfortunately this approach was going to result in two issues:
1) Chaining the filters themselves becomes a bit tricky, as the results
of the first filter would need to pass into the second. Probably doable
with an array of requestsWithPlatesFiltered that gets set and accessed
by index, but non-trivial.
2) We still couldn't apply the styling to the plates themselves without
significant complications as by theis point we've already filtered valid
plates. While we could build yet another computed from the post-filter
wells, its really over-engineering things at this stage.

So going to go back to the simple approach of checking the plates
themselves. But committing this branch in case we need to pursue this
route in future.